Sync up

+++ inner-loop-scripts/shell/deleteResourceGroups.sh
@@ -1,3 +1,7 @@
+#!/usr/bin/env bash
+
+set -e
+
 # This script might take about 30 minutes
 # Please check the variables
 RGLOCATION=$1
@@ -5,6 +9,37 @@ RGNAMEHUB=$2
 RGNAMESPOKES=$3
 RGNAMECLUSTER=$4
 AKS_CLUSTER_NAME=$5
+
+__usage="
+    [-c RGNAMECLUSTER]
+    [-h RGNAMEHUB]
+    [-k AKS_CLUSTER_NAME]
+    [-l LOCATION]
+    [-p RGNAMESPOKES]
+"
+
+usage() {
+    echo "usage: ${0##*/}"
+    echo "${__usage/[[:space:]]/}"
+    exit 1
+}
+
+while getopts "c:h:l:p:k:" opt; do
+    case $opt in
+    c)  RGNAMECLUSTER="${OPTARG}";;
+    h)  RGNAMEHUB="${OPTARG}";;
+    l)  LOCATION="${OPTARG}";;
+    p)  RGNAMESPOKES="${OPTARG}";;
+    k)  AKS_CLUSTER_NAME="${OPTARG}";;
+    *)  usage;;
+    esac
+done
+shift $(( $OPTIND - 1 ))
+
+if [ $OPTIND = 1 ]; then
+    usage
+    exit 0
+fi
 
 echo deleting $RGNAMECLUSTER
 az group delete -n $RGNAMECLUSTER --yes
